+++ google-chat-desktop/google-chat-desktop.csproj
@@ -42,6 +42,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <None Update="main\load\load.js">
+	    <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+	  </None>
 	  <None Update="Properties\Settings.settings">
 	    <Generator>SettingsSingleFileGenerator</Generator>
 	    <LastGenOutput>Settings.Designer.cs</LastGenOutput>
